fix: sgid update exclude and switch to use swapper
This was not excluding deq datasets that had their source as tables in SGID. So it was pulling the data from SGID during lift and then pushing it back to SGID during ship. To make matters worse, it was using Append which doesn't not have the ability to accept a geographic transformation so the data was being projected incorrectly.